Optimize integration tests where appropriate
Downloading builds or running the installer are not important for test
integrity.

+++ tests/integration/application/test_install.py
@@ -6,9 +6,15 @@ import zipfile
 import outputs
 import pytest
 
+import mons.commands.main
 from mons.mons import cli as mons_cli
 
 GITHUB_REPO = "https://github.com/EverestAPI/Everest"
+
+
+@pytest.fixture(autouse=True)
+def mock_installer(monkeypatch):
+    monkeypatch.setattr(mons.commands.main, "run_installer", lambda *args: True)
 
 
 def test_install_default(runner, test_install):
@@ -32,10 +38,11 @@ def test_install_url(runner, test_install):
     assert outputs.INSTALL_SUCCESS in result.output
 
 
-def test_install_zip(runner, test_install, tmp_path):
+def test_install_zip(runner, test_install, cache):
     url = f"{GITHUB_REPO}/releases/latest/download/olympus-build.zip"
-    file = os.path.join(tmp_path, "olympus-build.zip")
-    urllib.request.urlretrieve(url, file)
+    file = os.path.join(cache.mkdir("mons_test_install"), "olympus-build.zip")
+    if not os.path.exists(file):
+        urllib.request.urlretrieve(url, file)
 
     assert os.path.isfile(file)
 
@@ -49,14 +56,18 @@ def test_install_zip(runner, test_install, tmp_path):
     reason="no .NET build tool found",
     run=False,
 )
-def test_install_src(runner, test_install, tmp_path):
-    dest = os.path.join(tmp_path, "stable.zip")
-    urllib.request.urlretrieve(f"{GITHUB_REPO}/archive/refs/heads/stable.zip", dest)
+def test_install_src(runner, test_install, cache):
+    source_dir = os.path.join(cache.mkdir("mons_test_install"), "Everest-stable")
+    if not os.path.exists(source_dir):
+        dest = os.path.join(cache.mkdir("mons_test_install"), "stable.zip")
+        if not os.path.exists(dest):
+            urllib.request.urlretrieve(
+                f"{GITHUB_REPO}/archive/refs/heads/stable.zip", dest
+            )
 
-    with zipfile.ZipFile(dest, "r") as zip:
-        zip.extractall(tmp_path)
+        with zipfile.ZipFile(dest, "r") as zip:
+            zip.extractall(cache.mkdir("mons_test_install"))
 
-    source_dir = os.path.join(tmp_path, "Everest-stable")
     assert os.listdir(source_dir), "Failed to download source artifact: " + source_dir
 
     result = runner.invoke(mons_cli, ["install", test_install, "--src", source_dir])
